requirements.txt
pandas>=2.0.0
anthropic>=0.25.0
fastapi>=0.110.0
uvicorn>=0.29.0
python-dotenv>=1.0.0
pydantic>=2.0.0
